requirements.txt
numpy==1.11.2
pandas==0.19.0
py==1.4.31
pytest==3.0.3
python-dateutil==2.5.3
pytz==2016.7
PyYAML==3.12
requests==2.11.1
six==1.10.0
vcrpy==1.10.3
wrapt==1.10.8
